+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * update scala to 2.11.8
 * update "com.fasterxml.jackson.core" % "jackson-databind" to 2.7.3
 * update "joda-time" % "joda-time" to 2.9.3
+* `JNothing` is now treated the same way as `JNull` during the case-class serialization
 
 ## 0.6.2 (2016-01-07)